Backend/requirements.txt
# Flask Core & Web Framework
Flask==2.3.3
Werkzeug==2.3.7

# Environment Variables
python-dotenv==1.0.0

# Authentication & Security
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
flask-jwt-extended==4.5.3
PyJWT==2.8.0

# CORS & API Support
flask-cors==4.0.0

# Image Processing & Computer Vision
opencv-python-headless==4.8.1.78
Pillow==10.0.1

# OCR (Optical Character Recognition)
pytesseract==0.3.10

# Machine Learning & Embeddings (Lighter versions)
sentence-transformers==2.2.2
transformers>=4.35.2
torch>=2.1.0
numpy<2.0.0
scipy==1.10.1
scikit-learn==1.2.2

# PDF Generation
reportlab==4.0.4

# File Processing & Utilities
python-dateutil==2.8.2
requests==2.31.0
charset-normalizer==3.3.2
urllib3==2.0.7
idna==3.4
certifi==2023.7.22

# Add this to your requirements.txt
PyMuPDF==1.23.8
pymongo
flask_pymongo
bcrypt

flask==2.3.3
flask-jwt-extended==4.5.3
opencv-python==4.8.1.78
pillow==10.0.1
pytesseract==0.3.10
reportlab==4.0.6
sentence-transformers==2.2.2
numpy<2.0.0
requests==2.31.0
groq==0.4.1

# LangChain for RAG Pipeline (Chatbot)
langchain>=0.1.0
langchain-community>=0.0.20
langchain-groq>=1.0.0
chromadb>=0.4.0